refactor(waste-category): refactor many things
