+++ src/waste-category/aplication/usecases/delete.usecase.ts
@@ -1,9 +1,7 @@
 import type WasteCategoryRepository from '../../domain/repositories/waste-category.repository'
 
 class DeleteCategoryUseCase {
-  constructor(private readonly repository: WasteCategoryRepository) {
-    this.repository = repository
-  }
+  constructor(private readonly repository: WasteCategoryRepository) {}
 
   async exec(id: string): Promise<void> {
     await this.repository.delete(id)
