+++ src/waste-category/aplication/usecases/find-by-id.usecase.ts
@@ -2,10 +2,8 @@ import type WasteCategoryEntity from '../../domain/entities/waste-category.entit
 import ErrorCategoryNotFound from '../../domain/errors/category-not-found.error'
 import type WasteCategoryRepository from '../../domain/repositories/waste-category.repository'
 
-class FindCategoryByIDUseCase {
-  constructor(private readonly repository: WasteCategoryRepository) {
-    this.repository = repository
-  }
+class FindWasteCategoryByIDUseCase {
+  constructor(private readonly repository: WasteCategoryRepository) {}
 
   async exec(id: string): Promise<WasteCategoryEntity> {
     const categoryFound = await this.repository.find({ id })
@@ -17,4 +15,4 @@ class FindCategoryByIDUseCase {
   }
 }
 
-export default FindCategoryByIDUseCase
+export default FindWasteCategoryByIDUseCase
