+++ src/waste-category/aplication/usecases/list.usecase.ts
@@ -3,18 +3,12 @@ import ErrorCategoryNotFound from '../../domain/errors/category-not-found.error'
 import type WasteCategoryRepository from '../../domain/repositories/waste-category.repository'
 
 class ListCategoriesUseCase {
-  constructor(private readonly repository: WasteCategoryRepository) {
-    this.repository = repository
-  }
+  constructor(private readonly repository: WasteCategoryRepository) {}
 
   async exec(offset: number, limit: number): Promise<WasteCategoryEntity[]> {
     const entitiesFounded = await this.repository.list(offset, limit)
     if (entitiesFounded == null) {
-      throw new ErrorCategoryNotFound('Cannot find any category when try to list all waste categories')
-    }
-
-    if (entitiesFounded.length === 0) {
-      throw new ErrorCategoryNotFound('Cannot find any category when try to list all waste categories')
+      throw new ErrorCategoryNotFound()
     }
 
     return entitiesFounded
