+++ src/waste-category/infrastructure/dtos/update-waste-category.dto.ts
@@ -1,6 +1,5 @@
 import { z } from 'zod'
-import { descriptionDTO, nameDTO } from './dto-types/dto-types'
-import { pointsPerWeightDTO } from '../../../shared/infrastructure/dto-types/dto-types'
+import { nameDTO, pointsPerWeightDTO, descriptionDTO } from '../../../shared/infrastructure/dto-types/dto-types'
 
 const UpdateWasteCategoryDTO = z.object({
   name: nameDTO.optional(),
